Moved largest_connected_component_image from '_utils/ndimage.py' to 'image/image.py' and added it to 'image/__init__.py'. Moved tests for largest connected component from '/tests/test_ndimage' to '/image/tests/test_image.py'. Corrected some additional formatting errors. Added image.largest_connected_component to whats_new.rst
